Split out formatting tests by transformer
